move custom image logic to topcp codegen
