+++ code_generator_TopCPToolkit/servicex/TopCP_code_generator/request_translator.py
@@ -26,7 +26,9 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import os
+import json
 import shutil
+
 from . import query_translate
 from servicex_codegen.code_generator import (
     CodeGenerator,
@@ -34,6 +36,25 @@ from servicex_codegen.code_generator import (
     GenerateCodeException,
 )
 
+def validate_custom_docker_image(image_name: str) -> bool:
+    allowed_images_json = os.environ.get("TOPCP_ALLOWED_IMAGES")
+    print(allowed_images_json)
+
+    if not allowed_images_json:
+        raise GenerateCodeException("Custom Docker images are not allowed.")
+
+    try:
+        allowed_prefixes = json.loads(allowed_images_json)
+        if not isinstance(allowed_prefixes, list):
+            raise GenerateCodeException("TopCP allowed images are improperly configured.")
+        for prefix in allowed_prefixes:
+            if image_name.startswith(prefix):
+                return True
+
+        raise GenerateCodeException(f"Custom Docker image '{image_name}' not allowed.")
+
+    except json.JSONDecodeError:
+        raise GenerateCodeException("TopCP allowed images are improperly configured.")
 
 class TopCPTranslator(CodeGenerator):
     # Generate the code. Ignoring caching for now
@@ -63,11 +84,21 @@ class TopCPTranslator(CodeGenerator):
             "CAPABILITIES_PATH", "/home/servicex/transformer_capabilities.json"
         )
 
-        query_translate.generate_files_from_query(query, query_file_path)
+        jquery = json.loads(query)
+        query_translate.generate_files_from_query(jquery, query_file_path)
 
         shutil.copyfile(
             capabilities_path,
             os.path.join(query_file_path, "transformer_capabilities.json"),
         )
 
-        return GeneratedFileResult(_hash, query_file_path)
+        results = GeneratedFileResult(_hash, query_file_path)
+
+        if "docker_image" in jquery:
+            docker_image = jquery["docker_image"]
+
+            validate_custom_docker_image(docker_image)
+
+            results.image = docker_image
+
+        return results
